# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Hub customer API
  description: |
    A Flask + MongoDB API for the customer service.
    
    ## Authentication
    All endpoints (except `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a JWT from the Developer Edition welcome / IdP flow, or use a
    test token from `test/e2e/e2e_auth.py` that matches your `JWT_SECRET` and expected claims.
  version: 0.1.0
  contact:
    name: Mentor Forge
    email: info@agile-learning-institute.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8387
    description: Local development server
...
paths:
...
  /api/Card:
    get:
      tags:
        - Card
      summary: Get all Card documents
      description: Retrieve infinite scroll batch of sorted, filtered Card documents
      operationId: getCards
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Card
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Card
      summary: Create a new Card document
      description: Create a new Card document (requires staff or admin role)
      operationId: createCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardInput'
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Card/{CardId}:
    get:
      tags:
        - Card
      summary: Get Card by CardId
      description: Retrieve a specific Card document by its CardId
      operationId: getCard
      security:
        - bearerAuth: []
      parameters:
        - name: CardId
          in: path
          required: true
          description: The Card ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Card
      summary: Update Card document
      description: Update a Card document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateCard
      security:
        - bearerAuth: []
      parameters:
        - name: CardId
          in: path
          required: true
          description: The Card ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdate'
      responses:
        '200':
          description: Card updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/Customer:
    get:
      tags:
        - Customer
      summary: Get all Customer documents
      description: Retrieve infinite scroll batch of sorted, filtered Customer documents (read-only)
      operationId: getCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Customer
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved Customers
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/Customer/{CustomerId}:
    get:
      tags:
        - Customer
      summary: Get Customer by CustomerId
      description: Retrieve a specific Customer document by its CustomerId (read-only)
      operationId: getCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: CustomerId
          in: path
          required: true
          description: The Customer ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by your identity provider (Developer Edition welcome flow for local dev).
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
...
    Card:
      type: object
      description: Card domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Card name
          pattern: '^[^\s]{1,40}$'
          example: my-Card
        description:
          type: string
          description: Card description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Card document
        status:
          type: string
          description: Card status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    CardInput:
      type: object
      description: Input for creating a Card document
      properties:
        name:
          type: string
          description: Card name
          pattern: '^[^\s]{1,40}$'
          example: my-Card
        description:
          type: string
          description: Card description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Card document
        status:
          type: string
          description: Card status
          enum:
            - active
            - archived
          example: active
    CardUpdate:
      type: object
      description: Fields to update in a Card document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Card name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Card description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Card status
          enum:
            - active
            - archived
...
    Customer:
      type: object
      description: Customer domain document (read-only)
      additionalProperties: false
      required:
        - _id
        - name
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Customer name
          pattern: '^[^\s]{1,40}$'
          example: my-Customer
        description:
          type: string
          description: Customer description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Customer document
        status:
          type: string
          description: Customer status
          example: active
...